+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix: `HeaderNav.Button`: Add `cursor` to `onClick`. 
+
 ## [4.8.0] - 2021-09-22
 
 Features: