+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: `HeaderMenu`: Add `position` prop, default to `left`.
 - Fix: `HeaderNav.Button`: Add `cursor` to `onClick`. 
 
 ## [4.8.0] - 2021-09-22